+++ R/synchronousMultiplexSiteswap.R
@@ -228,7 +228,8 @@ method(throw_data, synchronousMultiplexSiteswap) <- function(
 method(timeline, synchronousMultiplexSiteswap) <- function(
   siteswap,
   n_cycles = 3,
-  title = TRUE
+  title = TRUE,
+  subtitle = TRUE
 ) {
   throw_data <- throw_data(siteswap, n_cycles = n_cycles)
 
@@ -246,11 +247,6 @@ method(timeline, synchronousMultiplexSiteswap) <- function(
       df
     }) |>
     purrr::list_rbind()
-
-  subtitle <- plot_subtitle(
-    siteswap,
-    extra = "Solid and dashed lines represent different hands."
-  )
 
   warn_if_props_hidden(siteswap, max_prop)
 
@@ -278,15 +274,18 @@ method(timeline, synchronousMultiplexSiteswap) <- function(
       axis.text.x = element_text(face = "bold", size = rel(1.5)),
       plot.margin = margin(10, 20, 20, 20)
     ) +
-    title_subtitle_theme()
-
-  if (title) {
-    p <- p +
-      labs(
-        title = paste0("Siteswap '", siteswap@sequence, "'"),
-        subtitle = subtitle
-      )
-  }
+    title_subtitle_theme() +
+    labs(
+      title = if (title) paste0("Siteswap '", siteswap@sequence, "'") else NULL,
+      subtitle = if (subtitle) {
+        plot_subtitle(
+          siteswap,
+          extra = "Solid and dashed lines represent different hands."
+        )
+      } else {
+        NULL
+      }
+    )
 
   p
 }
@@ -295,6 +294,7 @@ method(ladder, synchronousMultiplexSiteswap) <- function(
   siteswap,
   n_cycles = 3,
   direction = c("horizontal", "vertical"),
+  title = TRUE,
   subtitle = TRUE
 ) {
   direction <- rlang::arg_match(direction)
@@ -306,7 +306,11 @@ method(ladder, synchronousMultiplexSiteswap) <- function(
   build_ladder_plot(
     plot_data,
     direction,
-    paste0("Siteswap '", siteswap@full_sequence, "'"),
+    title = if (title) {
+      paste0("Siteswap '", siteswap@full_sequence, "'")
+    } else {
+      NULL
+    },
     subtitle = if (subtitle) plot_subtitle(siteswap) else NULL
   )
 }