+++ R/vanillaSiteswap.R
@@ -150,7 +150,8 @@ method(throw_data, vanillaSiteswap) <- function(siteswap, n_cycles = 3) {
 method(timeline, vanillaSiteswap) <- function(
   siteswap,
   n_cycles = 3,
-  title = TRUE
+  title = TRUE,
+  subtitle = TRUE
 ) {
   throw_data <- throw_data(siteswap, n_cycles = n_cycles)
 
@@ -166,8 +167,6 @@ method(timeline, vanillaSiteswap) <- function(
       generate_parabola(beat, catch_beat, throw, prop, beat)
     }) |>
     purrr::list_rbind()
-
-  subtitle <- plot_subtitle(siteswap)
 
   warn_if_props_hidden(siteswap, max_prop)
 
@@ -186,15 +185,11 @@ method(timeline, vanillaSiteswap) <- function(
       axis.text.x = element_text(face = "bold", size = rel(1.5)),
       plot.margin = margin(10, 20, 20, 20)
     ) +
-    title_subtitle_theme()
-
-  if (title) {
-    p <- p +
-      labs(
-        title = paste0("Siteswap '", siteswap@sequence, "'"),
-        subtitle = subtitle
-      )
-  }
+    title_subtitle_theme() +
+    labs(
+      title = if (title) paste0("Siteswap '", siteswap@sequence, "'") else NULL,
+      subtitle = if (subtitle) plot_subtitle(siteswap) else NULL
+    )
 
   p
 }
@@ -203,6 +198,7 @@ method(ladder, vanillaSiteswap) <- function(
   siteswap,
   n_cycles = 3,
   direction = c("horizontal", "vertical"),
+  title = TRUE,
   subtitle = TRUE
 ) {
   direction <- rlang::arg_match(direction)
@@ -214,7 +210,7 @@ method(ladder, vanillaSiteswap) <- function(
   build_ladder_plot(
     plot_data,
     direction,
-    paste0("Siteswap '", siteswap@sequence, "'"),
+    title = if (title) paste0("Siteswap '", siteswap@sequence, "'") else NULL,
     subtitle = if (subtitle) plot_subtitle(siteswap) else NULL
   )
 }